pyproject.toml
[tool.poetry]
name = "hw28"
version = "0.1.0"
description = ""
authors = ["mikepodolskiy <mikepodolskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
